UID and Path
Resources now store a uid and path for each sub-resource so that resources can be loaded in the editor as well as at run time. Eventually, a button will be added that when clicked in the editor, all stored paths will be updated by their UID in preparation for a build.