Slots and hide slots
Renaming a slot will now rename all hide slots in other slots to the new name.

The flow of renaming a slot in an AWOCSlotContainer has been reworked so that the confirm_overwrite dialog is displayed instead of the confirm_save dialog box, not in response to the confirm_save dialog box being confirmed.

The hide slot option button in AWOCSlotContainer will now set the selection to -1 so that nothing is selected when it is repopulated with options. This way, the add hide slot button can be set to active when an option in the hide slot option button is selected.

When a slot is renamed to a name that already exists and the confirm_overwrite dialog is confirmed, if the slot has a hide slot that is the same name as the slot, the hide slot will be removed.